MNT: move BlueskyEventStream to a separate module
